fix: offset computation bug
This PR fixes a bug when computing an offset from a previous match.
Before the fix, a previous match offset of 0 would be considered as
an out of bound offset and would cause the `Match::matches` function
to exit prematurely. This has been fix to consider 0 offset as a
valid offset.